can now pass params to router
